Move playlist to a different page
