+++ src/jsx/components/playlist/Entry.jsx
