--- src/less/components/playlist_app/playlist.less
@@ -1,88 +1,0 @@
-//
-// Dakara Project
-//
-// Playlist style file
-//
-
-
-// `playlist` id:
-//
-// This component stylizes the playlist, containing an info zone and the
-// playlist content.
-#playlist {
-
-    // The playlist info gives some pieces of information on the status of the
-    // playlist, tipycally the amount of songs in queue, the next song and when
-    // the playlist will end.
-    // Info don't have specific classes.
-    .playlist-summary {
-        cursor: pointer;
-        border: none;
-        outline: none;
-        width: 100%;
-        padding: 0 @gap-horizontal;
-        background: @brand-primary;
-        color: @body-background;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-
-        // restore fonts
-        .make-font;
-
-        &:hover,
-        &:focus, {
-            background: @brand-primary-lighter;
-        }
-
-        .item {
-            display: flex;
-            align-items: center;
-
-            .stat {
-                display: inline-block;
-                font-weight: 100;
-                font-size: 2.5em;
-                vertical-align: middle;
-            }
-
-            .description {
-                margin-left: 0.5em;
-                display: inline-block;
-                vertical-align: middle;
-                flex: 1;
-                text-align: left;
-            }
-        }
-    }
-
-    // The container for the playlist entries, based on `listing`, it can
-    // collapse.
-    .playlist-list {
-        @entries-list-height: @listing-entry-height * 5;
-        max-height: @entries-list-height;
-        overflow-y: auto;
-
-        // appearance animation
-        &.collapse-enter {
-            overflow: hidden;
-            max-height: 0;
-
-            &.collapse-enter-active {
-                max-height: @entries-list-height;
-                transition: max-height 300ms ease-out;
-            }
-        }
-
-        // disappearance animation
-        &.collapse-exit {
-            overflow: hidden;
-            max-height: @entries-list-height;
-
-            &.collapse-exit-active {
-                max-height: 0;
-                transition: max-height 150ms ease-out;
-            }
-        }
-    }
-}
