+++ src/less/components/playlist_app/playlist_app.less
@@ -5,6 +5,6 @@
 //
 
 
-@import "playlist";
+@import "playlist_info_bar";
 @import "player";
 @import "kara_status_notification";
