+++ src/less/components/playlist_app/playlist_info_bar.less
@@ -1,0 +1,48 @@
+//
+// Dakara Project
+//
+// Playlist info bar style file
+//
+
+
+// `playlist-info-bar` id:
+//
+// The playlist info bar gives some pieces of information on the status of the
+// playlist, tipycally the amount of songs in queue, the next song and when the
+// playlist will end.
+//
+// Info don't have specific classes.
+#playlist-info-bar {
+    cursor: pointer;
+    padding: 0 @gap-horizontal;
+    background: @brand-primary;
+    color: @body-background;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    &:hover,
+    &:focus, {
+        background: @brand-primary-lighter;
+    }
+
+    .item {
+        display: flex;
+        align-items: center;
+
+        .stat {
+            display: inline-block;
+            font-weight: 100;
+            font-size: 2.5em;
+            vertical-align: middle;
+        }
+
+        .description {
+            margin-left: 0.5em;
+            display: inline-block;
+            vertical-align: middle;
+            flex: 1;
+            text-align: left;
+        }
+    }
+}
